requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
aiohttp>=3.9.0
requests>=2.31.0
pyinstaller>=6.0.0
